Handle multiple license templates
Allows CMD properties to be defined as multivalued, in which case
values will be collected in an array instead of overwriting each
other. On output it will still throw away all but the last value,
but until then internal functions can access all of them.

The license matcher uses this feature to look through multiple
short names so multi-licensed images will be handled correctly if
any one of the licenses can be recongized.

Bug: 56687
Change-Id: If2732dd2eb7150c9ccfc4e82f4dcbf9a479a0fa9