Parse restrictions e.g. trademark, insignia, etc.
Find out what restrictions are imposed on a file that could have
legal ramifications, such as trademarks, insignia, nazi or
communist symbols, etc.  Takes elements with classes prefixed with
"restriction-" and outputs in extmetadata to "Restrictions".

Also, DataCollector::selectCoordinate() and selectDeletionReason()
have been replaced with selectFirst(), because they both did the
same thing.  The new function has an ambiguous name so that it can
be used for other things i.e. restrictions.

Bug: T77717
Change-Id: Ic7a4ec3a63a175dfc4f1d7907d0e81b7f831eb65