Add JSON-LD script with schema.org license info to filepages
Later this year Google will be adding a "licensable" tag to images
for which specific license metadata is available. This patch adds a
JSON-LD script to the end of File pages containing, if applicable,
a link to the image's license and a link to the File page, where
the user can read a summary of the license.

Bug: T254039
Change-Id: I5cdd92c62041c87252cca8988f03a24109a1f6e4